Store the software/cms and version in the platform, store the db password
This makes it easier to pass everything to Ansible as inventory variables.
While Aegir managed platform "packages" to handle variables, we generally
are not interested by all that, and only need to know what the CMS and
version is, so that we can support things like D7/D10, WordPress and
CiviCRM (without a CMS).

Storing the db pass will probably make a few people cringe, but Aegir
previously already stored in the db the root pass for MySQL, and the
drushrc.php files had permissions that would let www-data read the files.